dect: coa: fix preamble transmission
The first byte of the transmitted preamble is currently zero. For some
reason the firmware transmits 5 bytes of firmware, my assumption is that
the first byte is used during ramp-up of the radio.

According to ETS 175 300-02 Chapter 4.9, the ramp-up pattern must be
either the natural extension of the preamble pattern, or the value 0.
Both appear to work equally good, which seems to confirm my assumption.

Signed-off-by: Patrick McHardy <kaber@trash.net>