dect: dlc_cplane: fix mac connection leaks
Usually unbinding of the LC from the MAC connection and LC destruction
is triggered by the last LAPC unbinding from the LC. When no LAPCs are
bound to the LC, destruction is never triggered and the LC and MAC
connection leak.

Even worse, when trying to route the U-Plane to the MAC connection for
a call later on, the old disconnected MAC connection is found and used.

Fix by triggering LC destruction manually when no LAPCs are bound.

Signed-off-by: Patrick McHardy <kaber@trash.net>