dect: csf: synchronize timer when locking to FP
Fix synchronization to long running FPs: when the multiframe number is
above 2^23, the overflow-aware timer multiframe comparison will consider
timers meant to run immediately to be scheduled far in the future,
causing the channel list updates to fail, which in turn makes any
attempt to open a connection fail.

Roughly synchronize when locking to an FP to avoid this. Full synchronization
will be achieved once locked.

Signed-off-by: Patrick McHardy <kaber@trash.net>