dect: coa: reduce firmware size
Save a couple of instructions by moving the protected B-field transmission
and reception code directly into the calling functions. With this change all
receive and transmit functions call their respective B-field processing
functions in a similar fashion, allowing to remove the WT 1 for the last
A-field bit after Receive/Transmit, saving a total of 6 instructions.

Signed-off-by: Patrick McHardy <kaber@trash.net>