dect: fix scanning regression
Commit e783777 (dect: coa: write BMC configuration when setting
channel mode) moved writing the BMC configuration to the mode
setting callback.

The MAC CSF layer however didn't set the channel mode to SCANNING
when initiating a scan since BMC configuration was done during
initialization and mode setting in dect_transceiver_enable(). With
the above mentioned commit the BMC is never properly configured,
meaning the scan fails.

Properly set the mode of DECT_SCAN_SLOT to DECT_MODE_SCANNING when
initiating a scan to fix this.

Additionally get rid of the unnecessary BMC initialization when
initializing the DIP, this is done immediately afterwards anyways.

Signed-off-by: Patrick McHardy <kaber@trash.net>